extract types.ts
