+++ src/index.ts
@@ -17,7 +17,7 @@
 import RRule from './rrule'
 import RRuleSet from './rruleset'
 import RRuleStr from './rrulestr'
-export { Frequency, ByWeekday, WeekdayStr, Options } from './rrule'
+export { Frequency, ByWeekday, WeekdayStr, Options } from './types'
 export { default as Weekday } from './weekday'
 export { RRuleStrOptions } from './rrulestr'
