+++ src/rrule.ts
@@ -8,6 +8,7 @@ import CallbackIterResult from './callbackiterresult'
 import { Language } from './nlp/i18n'
 import { Nlp } from './nlp/index'
 import { GetText } from './nlp/totext'
+import { Cache, Days, ParsedOptions, Options, Frequency, CacheKeys } from './types'
 
 interface GetNlp {
   _nlp: Nlp
@@ -22,78 +23,9 @@ const getnlp: GetNlp = function () {
   return getnlp._nlp
 } as GetNlp
 
-export enum Frequency {
-  YEARLY = 0,
-  MONTHLY = 1,
-  WEEKLY = 2,
-  DAILY = 3,
-  HOURLY = 4,
-  MINUTELY = 5,
-  SECONDLY = 6
-}
-
 // =============================================================================
 // RRule
 // =============================================================================
-
-export interface Options {
-  freq: Frequency
-  dtstart: Date
-  interval: number
-  wkst: Weekday | number
-  count: number
-  until: Date
-  bysetpos: number | number[]
-  bymonth: number | number[]
-  bymonthday: number | number[]
-  bynmonthday: number[]
-  byyearday: number[]
-  byweekno: number | number[]
-  byweekday: ByWeekday | ByWeekday[]
-  bynweekday: number[][]
-  byhour: number | number[]
-  byminute: number | number[]
-  bysecond: number | number[]
-  byeaster: number
-}
-
-interface ParsedOptions {
-  freq: Frequency
-  dtstart: Date
-  interval: number
-  wkst: number
-  count: number
-  until: Date
-  bysetpos: number[]
-  bymonth: number[]
-  bymonthday: number[]
-  bynmonthday: number[]
-  byyearday: number[]
-  byweekno: number[]
-  byweekday: number[]
-  bynweekday: number[][]
-  byhour: number[]
-  byminute: number[]
-  bysecond: number[]
-  byeaster: number
-}
-
-type CacheKeys = 'before' | 'after' | 'between'
-type CacheBase = { [K in CacheKeys]: IterArgs[] }
-export type Cache = CacheBase & { all: Date[] | Partial<IterArgs>[] | false }
-
-const Days = {
-  MO: new Weekday(0),
-  TU: new Weekday(1),
-  WE: new Weekday(2),
-  TH: new Weekday(3),
-  FR: new Weekday(4),
-  SA: new Weekday(5),
-  SU: new Weekday(6)
-}
-
-export type WeekdayStr = keyof typeof Days
-export type ByWeekday = WeekdayStr | number | Weekday
 
 /**
  *
